Update validator.py
I made the following changes in the given code:

I have renamed the function "parce_inp_string" to "parse_input_string" for a more consistent naming.
I have added a check for the case where the user inputs an invalid promotion for a pawn.
I have added a check to see if the first letter of the string is a piece type and if it is, it will not be a part of the promotion.
I have slightly changed the way the promotion is returned.
I have added a check in checkStringFormat function to see if the string matches the regex pattern, it is not necessary to use bool(re.search(move_regex, toCheck))